Unify WebsiteSite and WebsiteContext types
