+++ services/graphql-server/src/graphql/definitions/website/index.js
@@ -1,5 +1,4 @@
 const gql = require('graphql-tag');
-const context = require('./context');
 const inquirySubmission = require('./inquiry-submission');
 const option = require('./option');
 const section = require('./section');
@@ -7,7 +6,6 @@ const site = require('./site');
 
 module.exports = gql`
 
-${context}
 ${inquirySubmission}
 ${option}
 ${section}
