+++ services/graphql-server/src/graphql/resolvers/website/site.js
@@ -1,5 +1,6 @@
 const querystring = require('querystring');
 const { UserInputError } = require('apollo-server-express');
+const defaults = require('../../defaults');
 
 const cleanRedirect = async (redirect, from, basedb) => {
   // Redirect already found. Do nothing.
@@ -14,6 +15,36 @@ const cleanRedirect = async (redirect, from, basedb) => {
 };
 
 module.exports = {
+  /**
+   *
+   */
+  WebsiteSite: {
+    origin: ({ host }) => `https://${host}`,
+    imageHost: ({ imageHost }) => imageHost || defaults.imageHost,
+    assetHost: ({ assetHost }) => assetHost || defaults.assetHost,
+    language: ({ language }) => ({ ...defaults.language, ...language }),
+    date: ({ date }) => ({ ...defaults.date, ...date }),
+  },
+
+  /**
+   *
+   */
+  WebsiteSiteLanguage: {
+    code: (language) => {
+      const { primaryCode, subCode } = language;
+      const primary = primaryCode.toLowerCase();
+      if (!subCode) return primary;
+      return `${primary}-${subCode.toLowerCase()}`;
+    },
+
+    primaryCode: language => language.primaryCode.toLowerCase(),
+
+    subCode: (language) => {
+      if (!language.subCode) return null;
+      return language.subCode.toLowerCase();
+    },
+  },
+
   /**
    *
    */
